+++ Assets/Scripts/ScriptableObjects/AttackTypesScrObj.cs
@@ -6,7 +6,7 @@ using System;
 [CreateAssetMenu(menuName = "Scriptable Objects/Attack Types")]
 public class AttackTypesScrObj : ScriptableObject
 {
-    public float damage;
+    public byte damage;
     public float knockback;
     public float playerPush;
     public Mesh colliderShape;
